Moving property endReview from uiSlide to root properties of slide
