+++ packages/@coorpacademy-components/src/organism/review-slide/test/fixtures/initial-state.js
@@ -1,11 +1,8 @@
 import map from 'lodash/fp/map';
 import set from 'lodash/fp/set';
 import AnswerQCMGraphic from '../../../../molecule/answer/test/fixtures/qcm-graphic';
-import AnswerQCMDrag from '../../../../molecule/answer/test/fixtures/qcm-drag';
 
 const markAnswersAsUnselected = answer => set('selected', false, answer);
-
-const qcmDrag = AnswerQCMDrag.props;
 const qcmGraphic = AnswerQCMGraphic.props;
 qcmGraphic.model.answers = map(markAnswersAsUnselected, AnswerQCMGraphic.props.model.answers);
 
@@ -22,23 +19,18 @@ export default {
       },
       '1': {
         hidden: false,
-        position: 1,
-        questionText: 'Question 2', // c'est pas censé d'afficher ça dans la slide derrière ?
-        answerUI: qcmDrag
+        position: 1
       },
       '2': {
         hidden: false,
-        questionText: 'Question 3',
         position: 2
       },
       '3': {
         hidden: false,
-        questionText: 'Question 4',
         position: 3
       },
       '4': {
         hidden: false,
-        questionText: 'Question 5',
         position: 4
       }
     },
