+++ packages/@coorpacademy-components/src/organism/review-stacked-slides/test/fixtures/end-review.js
@@ -1,78 +1,8 @@
-import AnswerQCMDrag from '../../../../molecule/answer/test/fixtures/qcm-drag';
+import cloneDeep from 'lodash/fp/cloneDeep';
+import AllOkFixture from './all-ok';
 
-const qcmDrag = AnswerQCMDrag.props;
+const EndReviewFixture = cloneDeep(AllOkFixture);
+EndReviewFixture.props.uiSlides['4'].animateCorrectionPopin = false;
+EndReviewFixture.props.endReview = true;
 
-export const correctionPopinProps = {
-  klf: {
-    label: 'Key learning factor',
-    tooltip: 'Some tooltip info.'
-  },
-  information: {
-    label: 'Key learning factor',
-    message: 'info msg'
-  },
-  next: {
-    label: 'Next',
-    onClick: () => {
-      console.log('Load next slide');
-    }
-  },
-  type: 'right',
-  resultLabel: 'CORRECT ANSWER'
-};
-
-export default {
-  props: {
-    uiSlides: {
-      '0': {
-        hidden: false,
-        position: 0,
-        questionText: 'Hey there, .....suspense.... ready to select some answers?',
-        answerUI: qcmDrag,
-        isCorrect: true,
-        endReview: true,
-        showCorrectionPopin: true
-      },
-      '1': {
-        hidden: true,
-        questionText: 'Question 2',
-        position: 1,
-        endReview: true
-      },
-      '2': {
-        hidden: true,
-        questionText: 'Question 3',
-        position: 2,
-        endReview: true
-      },
-      '3': {
-        hidden: true,
-        questionText: 'Question 4',
-        position: 3,
-        endReview: true
-      },
-      '4': {
-        questionText: 'Question 5',
-        hidden: true,
-        position: 4,
-        endReview: true
-      }
-    },
-    validateButton: {
-      label: 'Validate',
-      onClick: () => console.log('onValidateClick'),
-      disabled: true
-    },
-    finishedSlides: {
-      // Incohérent pour la fixture qui fait la fin de toutes les slides
-      '0': true
-    },
-    finishedSlidesSize: 1, // Incohérent
-    correctionPopinProps,
-    updateSlidesOnNext: () => console.log('updateSlidesOnNext'), // à revoir
-    updateStepItemsOnNext: () => console.log('updateStepItemsOnNext'), // à revoir, on devrait avoir un seul onClick dans le next
-    updateReviewStatus: () => console.log('updateReviewStatus'),
-    updateStepItemsOnValidation: () => console.log('updateStepItemsOnValidation'),
-    updateFinishedSlides: () => console.log('updateFinishedSlides')
-  }
-};
+export default EndReviewFixture;
