+++ packages/@coorpacademy-components/src/organism/review-stacked-slides/test/fixtures/first-correct.js
