+++ packages/@coorpacademy-components/src/organism/review-stacked-slides/test/fixtures/initial-state.js
@@ -1,5 +1,5 @@
 import omit from 'lodash/fp/omit';
-import initStateFixture from '../../../review-slide/test/fixtures/default';
+import initStateFixture from '../../../review-slide/test/fixtures/initial-state';
 
 export default {
   props: {
