+++ packages/@coorpacademy-components/src/template/app-review/player/test/review-player.js
